requirements.txt
chainlit
openai
langsmith
mysql-connector-python
tabulate
python-decouple